requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
comm==0.1.3
cryptography==41.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.5
django-allauth==0.54.0
django-cors-headers==4.1.0
django-filter==23.2
django-import-export==3.2.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
et-xmlfile==1.1.0
executing==1.2.0
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
MarkupPy==1.14
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.1
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.8.0.74
openpyxl==3.1.2
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
stripe==5.4.0
tablib==3.5.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
zipp==3.16.2
